requirements.txt

fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
pyyaml==6.0.2
ipaddress==1.0.23
